Move data types from `remote` to `types`.
